Split country utils specs
