+++ test/unit/app/countries/utils.spec.js
@@ -23,50 +23,6 @@ import { getSortedCountryListFromProposals } from '../../../../src/app/countries
 import { QualityLevel } from 'mysterium-vpn-js'
 
 describe('countries utils', () => {
-  describe('.getSortedCountryListFromProposals', () => {
-    const proposals = [
-      {
-        providerId: '0x1234567890',
-        serviceDefinition: {
-          locationOriginate: {
-            country: 'LT'
-          }
-        }
-      },
-      {
-        providerId: '0x9876543210',
-        serviceDefinition: {
-          locationOriginate: {
-            country: 'AU'
-          }
-        }
-      },
-      {
-        providerId: '0x0987654321',
-        serviceDefinition: {
-          locationOriginate: {
-            country: 'CD'
-          }
-        }
-      }
-    ]
-
-    it('returns sorted list', () => {
-      const list = getSortedCountryListFromProposals(proposals, new Set(['0x0987654321']))
-      expect(list[0].id).to.be.eql('0x0987654321')
-      expect(list[0].name).to.be.eql('Congo, The Democratic Republic of the')
-      expect(list[0].code).to.be.eql('CD')
-
-      expect(list[1].id).to.be.eql('0x9876543210')
-      expect(list[1].name).to.be.eql('Australia')
-      expect(list[1].code).to.be.eql('AU')
-
-      expect(list[2].id).to.be.eql('0x1234567890')
-      expect(list[2].name).to.be.eql('Lithuania')
-      expect(list[2].code).to.be.eql('LT')
-    })
-  })
-
   describe('.getCountryLabel', () => {
     const proposals = [
       {
